Create GLES2Showroom.zgeproj
Here's the current state of my GLES2 showroom. It's great to show how very simple ES2/GL3 shaders work and exchange needed data with ZGE.
I hope it will help newcomers that are starting to use this OpenGL mode in ZGE, as it is NOT simple.
Don't hesitate to add other examples to it!